Net settlement group reject tests implemented
