+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/CreateMeteringPoints/ProductionMeteringPointTests.cs
@@ -114,7 +114,7 @@ namespace Energinet.DataHub.MeteringPoints.IntegrationTests.CreateMeteringPoints
 
             await SendCommandAsync(request).ConfigureAwait(false);
 
-            AssertValidationError("D02");
+            AssertValidationError("D62");
         }
 
         [Fact]
